+++ lib/src/buttons/popup_button.dart
@@ -1194,10 +1194,13 @@ class _MacosPopupButtonState<T> extends State<MacosPopupButton<T>>
       }
 
       hintIndex = items.length;
-      items.add(IgnorePointer(
-        ignoringSemantics: false,
-        child: displayedHint,
-      ));
+      items.add(
+        ExcludeSemantics(
+          child: IgnorePointer(
+            child: displayedHint,
+          ),
+        ),
+      );
     }
 
     // If value is null (then _selectedIndex is null) then we