+++ lib/src/enums/accent_color.dart
@@ -1,0 +1,27 @@
+/// The macOS accent color which can be changed by the user in *System Settings*
+/// → *Appearance* → *Accent color*.
+enum AccentColor {
+  /// The blue accent color.
+  blue,
+
+  /// The purple accent color.
+  purple,
+
+  /// The pink accent color.
+  pink,
+
+  /// The red accent color.
+  red,
+
+  /// The orange accent color.
+  orange,
+
+  /// The yellow accent color.
+  yellow,
+
+  /// The green accent color.
+  green,
+
+  /// The graphite accent color.
+  graphite,
+}