+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [2.0.1]
+### 🔄 Updated 🔄
+* `PushButton` has received a facelift. It now mimics the look and feel of native macOS buttons more closely.
+  * **Note:** As a result, its `pressedOpacity` property and the `PushButtonTheme` class have been deprecated. 
+
 ## [2.0.0]
 ### 🚨 Breaking Changes 🚨
 * `macos_ui` has been migrated to utilize [macos_window_utils](https://pub.dev/packages/macos_window_utils) under the hood, which provides the following benefits: